+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+* Add support for Fabric Native Components in AdView.
 * Add support for Fabric Native Components in NativeAdView.
 * Add support for Turbo Native Modules.
 ## 8.2.0